Isolate Timedelta scalar tests
